Remove traditional auth and getAuthParams
