+++ vscode/src/utils/links.ts
@@ -55,7 +55,7 @@ export const getLinks = async (globalState: GlobalState): Promise<Link[]> => {
     const gitOrg = globalState.getGitOrg();
     try {
         const codesResponse = await axios.get(`${API_ENDPOINT}/links`, {
-            params: { userId, subdomain, repo, gitOrg }
+            params: { ...globalState.getAuthParams(), repo, gitOrg }
         });
         return codesResponse.data.codes;
     } catch (err) {
@@ -73,10 +73,7 @@ export const deleteLink = async (globalState: GlobalState, linkId?: string): Pro
     const userId = globalState.getUserId();
     if (subdomain == null || userId == null) {return;} // TODO - proper error handling
     await axios.delete(`${API_ENDPOINT}/links/${linkId}`, {
-        params: {
-            userId,
-            subdomain,
-        }
+        params: globalState.getAuthParams()
     });
 };
 
@@ -85,10 +82,7 @@ export const deleteDoc = async (globalState: GlobalState, docId: string): Promis
     const userId = globalState.getUserId();
     if (subdomain == null || userId == null) {return;}
     await axios.delete(`${API_ENDPOINT}/docs/${docId}`, {
-        params: {
-            userId,
-            subdomain,
-        }
+        params: globalState.getAuthParams()
     });
 };
 
@@ -97,9 +91,6 @@ export const editDocName = async (globalState: GlobalState, docId: string, newNa
     const userId = globalState.getUserId();
     if (subdomain == null || userId == null) {return;}
     await axios.put(`${API_ENDPOINT}/docs/${docId}/title`, { title: newName }, {
-        params: {
-            userId,
-            subdomain,
-        },
+        params: globalState.getAuthParams(),
     });
 };
